replace patches of WrapMusic-reliant scripts with compiled ones and clean up
Since commit f960ce5 we don't need to hex edit the patches of the scripts that rely on WrapMusic, I'm replacing them with compiled counterparts. Enabled unused code in rm310 to not require a heap patch, disabled part of it to avoid compilation errors. Disabled part of the unused code in UnscrewBulb as well, it isn't needed. SCICompanion altered rm710.sco when compiling rm710 without a good reason, matches the one we had right before commit 00c15c2 and proves to be inocuous.