fix wrong music triggering in rm520
During act 4 after the loose cobra scene, if you left the room without examining the countess' body and reentered, the 'loose cobra music' was wrongly triggered again. Flag 80 is set when the player puts the cobra back into the jail, so we fix it by adding flag 80 to the group of checks used to decide what music to play.