prevent stack overflow in ArmorFlag (448)
The game can crash or freeze in ArmorFlag (448) during the act 5 chase. The sHeKills script is set as the next room script when pursuitTimer expires, via rm448:notify, not checking if the current room script is already sHeKills. On the other hand, if the player returns to ArmorFlag (448) without hiding in the mummy coffin in 454 and without having opened the transom, the global pursuitTimer is cued to run the sHeKills script, but this doesn't stop the timer. If pursuitTimer expires during sHeKills, rm448:notify will schedule sHekills, leaving sHeKills as both the current and the next room script and causing infinite recursion when changing rooms. Fixed by adding a test in rm448:notify to not schedule sHeKills if it's the current script. Reference: https://github.com/scummvm/scummvm/blob/85702e06764f95a6b700e348dd90931613efdc29/engines/sci/engine/script_patches.cpp#L12035