fix transom re-initialization in ArmorFlag (448)
During the act 5 chase, room 448 allows moving the chair and opening/closing the transom. Flag 116 is set when the chair is moved and flag 39 is set when the transom is opened. But on rm448:init, the cel of the transom is decided by checking if either flag is set, making it appear open when the player only moved the chair. Fixed by independently using flag 116 for the chair and flag 39 for the transom. Reference: https://github.com/scummvm/scummvm/blob/85702e06764f95a6b700e348dd90931613efdc29/engines/sci/engine/script_patches.cpp#L12109