rm454 ScummVM compatibility fix
ScummVM's SCI engine script_patches.cpp disables seven of the eight '(gIconBar disable: 7)' calls in rm454. Creating a DOSBox patch to replicate this would cause a conflict where the eighth instance would also get disabled if our patch was to be used under ScummVM. To avoid this, we modify the bytecode signature of the eighth instance by changing it to (gIconBar disable: (+ local1 6)). Since the previous instruction sets local1 to 1, we know (+ local1 6) will always evaluate to 7 in this context, solving the ScummVM conflict without introducing new issues.