correctly revert accidental change in rm350
Due to a mistake, commit 2acdf76 restored rm350.sc to the state it had in 95f144f, instead of restoring it to 03d6b8f. This commit correctly restores the bugfix we had in 03d6b8f and improves on it, placing '(gGame handsOff:)' before '(proc0_3 25)', as it makes way more sense to disable user input right at the start of sEnterSouth:changeState. Tweaked some of the comments and recompiled 350.scr to reflect the change.