fix eastern exit lockup in rm460
If the player enters room 460 from the elevator room (660) and swings the hanging crate, the game softlocks when trying to go through the eastern door. Moving the crate that reveals the door sets local0 to 1, and sSwingIt:changeState checks if local0 is set to decide which polygon should it use as the room's obstacle. The problem is that local0 it isn't re-initialized after re-entering the room, causing sSwingIt:changeState to set the wrong polygon. We fix it by making sSwingIt:changeState check if the previous room was 660 and set local0 to 1 if that's the case, before it does the check to decide which polygon to use. As described in: https://github.com/scummvm/scummvm/blob/85702e06764f95a6b700e348dd90931613efdc29/engines/sci/engine/script_patches.cpp#L11747