fix hands-off cursor being displayed when player has control (hex editing involved)
When speech is enabled, the game often shows the hands-off cursor when the player has control. The Narrator records and restores the cursor being used, and if handsOn is called after say, the Narrator restores the hands-off cursor when the speech completes. Calling handsOn before say doesn't case this problem, but many scripts call these in the wrong order. This fix makes handsOn detect if a Narrator is saying a message, updating its saveCursor to the correct cursor when that's the case. Fixed as described in: https://github.com/scummvm/scummvm/blob/85702e06764f95a6b700e348dd90931613efdc29/engines/sci/engine/script_patches.cpp#L11279