Corrections related to the closing painting/safe bugfix in rm560 and DialFeature
Fixed a typo in rm560's comment. Adjusted the modNum that makes safePicture impersonate safePic, in DialFeature (script #561). Using 560 worked perfectly fine but is redundant, since safePicture (and everything else in script #561) is always used while in room 560. Using -1 (the default) is equivalent of using the current room number for the modNum.