Simplify rm700's visual bugfix and document code. Add missing ';' to rm310.
rm700's fix to correct the glitchy visual remains can be simplified, as '(Animate gEgo)' was invalid because it isn't cast, and it worked because it's equivalent to using '(Animate 0)' or '(Animate)'. Added one missing ';' to one of rm310's comments, that doesn't affect code at all but is an oversight.