Remove fake memory limitation to show the about screen, in Main (#0)
The game lies to the player telling there isn't enough memory to show the about screen, when what's actually being checked is if the current act is act 2 or the current room is in a list of rooms. We disable that restriction by hex editing 0.scr at 0x0DEE changing '89 7B' (lsg global123) to '33 5A' (jmp 90 bytes).